Refactor bounding box and point models
