--- app/models/label/instance/point.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-module Label
-  class Instance
-    # Tableless
-    # Label instance point
-    class Point
-      attr_accessor :top, :left
-
-      def initialize(top, left)
-        @top = top
-        @left = left
-      end
-    end
-  end
-end
